Fix for reserved keywords in struct members
Fixes previous problem where JSON variables where not able to be aliased to D members of a struct if they had the same name as a reserved D keyword. For example, "module" and "version" are now aliased to "_version" and "_module" for struct and class members. This does not affect associative arrays.